refactor(frontend): split task read flow
